Fixes #3624: stop truncating CDX root/free-node pointers to 16 bits
parse_cdx_header() read the VFP compound-index (.CDX/.DCX) root-node and
free-node-list pointers as 16-bit words, but these are documented 32-bit
LE fields. This truncates the pointers for any real CDX index whose root
node lives past byte offset 65535 (routine once an index grows past
~128 pages), either rejecting a valid file outright or reporting a bogus
offset to diagnostic tooling. The same file already defines and uses a
32-bit LE reader elsewhere in the same function for tag-page offsets, so
this was a wrong-width reader rather than a deliberate simplification.

Widened CdxHeader::root_node_offset/next_free_node_offset to uint32_t
(matching the already-uint32_t IndexProbeResult::free_node_offset_hint)
and read them via the existing 32-bit reader.

+++ include/copperfin/vfp/cdx_header.h
@@ -13,8 +13,8 @@ namespace copperfin::vfp {
 
 struct CdxHeader {
     std::array<std::uint16_t, 8> raw_words{};
-    std::uint16_t root_node_offset = 0;
-    std::uint16_t next_free_node_offset = 0;
+    std::uint32_t root_node_offset = 0;
+    std::uint32_t next_free_node_offset = 0;
     std::uint16_t key_length_hint = 0;
     std::uint16_t key_pool_length_hint = 0;
     std::uint16_t page_size = 512;
